+++ tracing/jaeger/config.go
@@ -17,3 +17,10 @@ func NewConfig(serviceName string) *Config {
 		ServiceName: serviceName,
 	}
 }
+
+// AppConfig can be used in an application config to represent Jaeger connection details.
+// It supports github.com/goph/nest
+type AppConfig struct {
+	Enabled bool   `env:""`
+	Addr    string `env:""`
+}